FIX remove pytorch solver not ready
