+++ solvers/sklearn.py
@@ -1,0 +1,47 @@
+import warnings
+
+
+from benchopt import BaseSolver
+from benchopt import safe_import_context
+
+
+with safe_import_context() as import_ctx:
+    from sklearn.exceptions import ConvergenceWarning
+    from sklearn.linear_model import LogisticRegression
+    from scipy.optimize.linesearch import LineSearchWarning
+
+
+class Solver(BaseSolver):
+    name = 'sklearn'
+
+    install_cmd = 'conda'
+    requirements = ['scikit-learn']
+
+    parameters = {
+        'solver': [
+            'liblinear',
+            'newton-cg',
+            'lbfgs',
+        ],
+    }
+    parameter_template = "{solver}"
+
+    def set_objective(self, X, y, lmbd):
+        self.X, self.y, self.lmbd = X, y, lmbd
+
+        warnings.filterwarnings('ignore', category=ConvergenceWarning)
+        warnings.filterwarnings('ignore', category=LineSearchWarning)
+        warnings.filterwarnings('ignore', category=UserWarning,
+                                message='Line Search failed')
+
+        self.clf = LogisticRegression(
+            solver=self.solver, C=1 / self.lmbd,
+            penalty='l2', fit_intercept=False, tol=1e-15
+        )
+
+    def run(self, n_iter):
+        self.clf.max_iter = n_iter
+        self.clf.fit(self.X, self.y)
+
+    def get_result(self):
+        return self.clf.coef_.flatten()
